pyproject.toml
[tool.poetry]
name = "repl_python3_Birthday"
version = "0.1.0"
description = "Always On Repl for Birthday Bot"
authors = ["Shourya De <shourya.de12@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
flask = "^2.0.1"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
